requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
ajsonrpc==1.2.0
alabaster==1.0.0
altgraph==0.17.4
anyio==4.4.0
appdirs==1.4.4
Arpeggio==2.0.2
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
babel==2.16.0
blinker==1.9.0
bottle==0.12.25
category_encoders==2.7.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
depthai-viewer==0.1.7
distlib==0.3.9
dnspython==2.7.0
docutils==0.21.2
eventlet==0.33.3
exceptiongroup==1.2.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
Flask-SocketIO==4.1.0
fonttools==4.54.1
frozenlist==1.5.0
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.0
h11==0.14.0
idna==3.10
imagesize==1.4.1
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
iniconfig==2.0.0
interchange==2021.0.4
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.3.2
jsonpickle==3.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
lightgbm==4.5.0
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
myst-parser==4.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.0
numpy==1.26.4
odrive==0.6.9.post0
onnx==1.17.0
onnxconverter-common==1.14.0
opencv-contrib-python==4.11.0.86
orjson==3.10.15
packaging==24.0
paho-mqtt==2.1.0
pandas==2.1.4
pansi==2024.11.0
parso==0.8.4
path==17.0.0
patsy==1.0.1
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.3.6
platformio==6.1.15
plotly==5.24.1
plotly-resampler==0.10.0
pluggy==1.4.0
pmdarima==2.0.4
portion==2.6.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==3.20.2
psutil==6.1.1
pure_eval==0.2.3
py2neo==2021.2.4
pyarrow==10.0.1
pycaret==3.3.2
pycparser==2.22
pyelftools==0.31
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.8
pyod==2.0.3
pyparsing==3.2.0
pyserial==3.5
pytest==8.0.1
python-dateutil==2.9.0.post0
python-engineio==3.13.0
python-socketio==4.2.0
pytz==2025.1
pyusb==1.3.1
pywin32==308
pywin32-ctypes==0.2.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.22.3
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
semantic-version==2.10.0
six==1.16.0
skl2onnx==1.18.0
sklearn-compat==0.1.3
sktime==0.26.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==8.1.3
sphinx-markdown-tables==0.0.17
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.4
tabulate==0.9.0
tbats==1.1.3
tenacity==9.0.0
termcolor==2.5.0
textX==4.1.0
threadpoolctl==3.5.0
tomli==2.0.2
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typing_extensions==4.12.1
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.29.0
virtualenv==20.27.1
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
yellowbrick==1.5
zipp==3.21.0
zope.event==5.0
zope.interface==7.2